fix: change the implementation of SymblAI API
The changes were made to have more modular code. The changes
include the creation of a new class ApiLinkViewModel which
purpose is to pass the analysed data across classes, and a
big refactoring of the SymblApiClient.kt file. This class now
checks when the job is done using the corresponding SymblAI
feature, instead of polling until the message is ready.