enhancement: add offline prompts repository Files created: OfflinePromptsRepoInterface, OfflinePromptsRepository - Files in which I have implemented helper functions to make testing easier later on. They also contain new functions which will allow to write the feedback to files later on
Modified files:
MainActivity
ChatViewModel - to include clearer Log statements for future debugging (which will be helpful when writing the prompts to the file). The function offlineRequest now takes in a new parameter (interviewID) which will be necessary to write the prompt in a file.
OfflineInterviewModule, OfflineRecordingsProfile, PreviousRecordingsFeedback - to include the new repository and use its functions instead of OfflinePromptsFunctionsImplementation

Files deleted:
OfflinePromptsFunctionsImplementation - Old file in which the helper functions were located, it's not a UI file so it was in the wrong place. Its old contents are now located in OfflinePromptsRepository.